+++ src/audio/SoundCache.h
